+++ packages/in_app_purchases_interface/lib/interface/callback_interface.dart
@@ -1,6 +1,8 @@
 import 'package:in_app_purchases_interface/in_app_purchases_interface.dart';
 
 abstract class CallbackInterface {
+  /// Restore the Purchase
   Future<bool> restore();
+  // Make a Purchase
   Future<bool> purchase(SubscriptionData productDetails);
 }